requirements.txt
pathlib
langchain_community
langchain
langchain_google_genai
python-dotenv
pypdf
streamlit
reportlab
docx2txt
python-docx
openai